+++ Source/DataSources/CylinderGraphics.js
@@ -35,8 +35,7 @@ define([
      * @param {Property} [options.outlineColor=Color.BLACK] A Property specifying the {@link Color} of the outline.
      * @param {Property} [options.outlineWidth=1.0] A numeric Property specifying the width of the outline.
      * @param {Property} [options.numberOfVerticalLines=16] A numeric Property specifying the number of vertical lines to draw along the perimeter for the outline.
-     * @param {Property} [options.slices=128] The number of edges around perimeter of the cylinder.
-     *
+     * @param {Property} [options.slices=128] The number of edges around the perimeter of the cylinder.
      */
     function CylinderGraphics(options) {
         this._length = undefined;
@@ -103,7 +102,6 @@ define([
 
         /**
          * Gets or sets the Property specifying the number of vertical lines to draw along the perimeter for the outline.
-         * to use when outlining the cylinder.
          * @memberof CylinderGraphics.prototype
          * @type {Property}
          * @default 16
@@ -111,8 +109,7 @@ define([
         numberOfVerticalLines : createPropertyDescriptor('numberOfVerticalLines'),
 
         /**
-         * Gets or sets the Property specifying the number of edges around perimeter of the cylinder.
-         * between each latitude and longitude point.
+         * Gets or sets the Property specifying the number of edges around the perimeter of the cylinder.
          * @memberof CylinderGraphics.prototype
          * @type {Property}
          * @default 16